fix: method moving
